[GlobalISel] Handle multiple registers in dbg.value intrinsic
https://reviews.llvm.org/D66077

This is a minor variation of the patch which was upstreamed in D66077
for the change in the return type of getOrCreateVRegs.

The value passed into dbg.value may relate to multiple registers,
each of which need a DBG_VALUE.

This fix calls MIRBuilder.buildDirectDbgValue for each register.

Without this, IR passed in from flang-compiler/flang may fail an
assertion in getOrCreateVReg.

Patch by : peterwaller-arm.

llvm-svn: 369403
Signed-off-by: Peter Waller <peter.waller@arm.com>